feat: 일정 추가 버튼 일정 갯수에 따른 위치 변하는 점 수정

+++ app/src/main/java/com/hifi/redeal/schedule/model/model.kt
@@ -1,4 +1,6 @@
 package com.hifi.redeal.schedule.model
 
-data class ScheduleData(val clientIdx: Int, val isScheduleFinish: Boolean, val isVisitSchedule: Boolean, val scheduleContext: String,
+import com.google.firebase.Timestamp
+
+data class ScheduleData(val clientIdx: Long, val isScheduleFinish: Boolean, val isVisitSchedule: Boolean, val scheduleContext: String,
 val scheduleDataCreateTime: String, val scheduleDeadlineTime: String, val scheduleTitle: String)
